+++ index.js
@@ -1,10 +1,11 @@
 const express = require("express");
 const exphbs = require("express-handlebars");
 const usersRouter = require("./routes/users");
+const resenhasRouter = require("./routes/resenhas");
 const painelRouter = require("./routes/painel");
 const dotenv = require("dotenv");
 const connectToDatabase = require("./database/connect");
-const { model, models } = require("mongoose");
+// const { model, models } = require("mongoose");
 const app = express();
 const hbs = exphbs.create({ partialsDir: ["views/partials"] });
 const port = process.env.PORT || 8080;
@@ -18,6 +19,7 @@ dotenv.config();
 connectToDatabase();
 
 app.use("/user", usersRouter);
+app.use("/resenhas", resenhasRouter);
 app.use("/painel", painelRouter);
 
 app.get("/", (req, res) => {
