+++ views/home.handlebars
@@ -4,10 +4,6 @@
       alt="Capa do livro, jogos vorazes 1"
     /></figure>
   <figure><img
-      src="/img/sertoes-veredas.jpg"
-      alt="Capa do livro, jogos vorazes 2"
-    /></figure>
-  <figure><img
       src="/img/oblomov.jpg"
       alt="Capa do livro, jogos vorazes 2"
     /></figure>
